Remove duplicate nodes from tree
Problem:
If user provides two page IDs to backup and one page being part/child of
another, then the child page is added to root node as well as to another page
as a child.
This creates a duplicate node as well as disrupts the tree structure which does
not match the hierarchy of pages present on Notion website.
Similar problem occurs if database is indirect/direct part of another page or
vice-versa

Solution:
Check if provided page/database ID to backup is part of some another page and
same page is also being backed up, then don't add such page/database as a part
of root node.